backend/requirements.txt
# SliverUI Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0

# Sliver Client
sliver-py>=0.0.19

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
aiosqlite>=0.19.0

# Redis
redis>=5.0.1
celery>=5.3.6

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Async
httpx>=0.26.0
aiofiles>=23.2.1

# Utilities
python-dateutil>=2.8.2
